move component js and scss into sub folders
