+++ my-app/src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import Header from './components/Header';
-import Footer from './components/Footer';
-import './App.css';
+import Header from './components/Header/Header';
+import Footer from './components/Footer/Footer';
+import './style.css';
 
 class App extends Component {
   render() {
