+++ my-app/src/components/Footer/Footer.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
+import './footer.css';
 
 class Footer extends Component {
   render() {
-    return <footer>Footer</footer>;
+    return <footer className="footer">Footer</footer>;
   }
 }
