+++ my-app/src/components/Footer/footer.scss
@@ -1,0 +1,8 @@
+@import 'styles/_variables.scss';
+
+.footer {
+	width: 100%;
+	display: flex;
+	color: $black;
+	background: red;
+}
