+++ my-app/src/components/Header/Header.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
+import './header.css';
 
 class Header extends Component {
   render() {
-    return <header>header</header>;
+    return <header className="header">header</header>;
   }
 }
