+++ my-app/src/components/Header/header.scss
@@ -1,0 +1,8 @@
+@import 'styles/_variables.scss';
+
+.header {
+	width: 100%;
+	display: flex;
+	background: $black;
+	color: #fff;
+}
