+++ my-app/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './App.css';
+import './style.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
